+++ examples/snippets/logger/correlationIdLogger.ts
@@ -1,0 +1,6 @@
+import { Logger } from '@aws-lambda-powertools/logger';
+import { search } from '@aws-lambda-powertools/logger/correlationId';
+
+const logger = new Logger({
+  correlationIdSearchFn: search,
+});