+++ packages/logger/src/index.ts
@@ -1,4 +1,4 @@
-export { Logger } from './Logger.js';
+export { LogLevel, LogLevelThreshold } from './constants.js';
 export { LogFormatter } from './formatter/LogFormatter.js';
 export { LogItem } from './formatter/LogItem.js';
-export { LogLevel, LogLevelThreshold } from './constants.js';
+export { Logger } from './Logger.js';