+++ packages/logger/src/logBuffer.ts
@@ -1,5 +1,10 @@
 import { isString } from '@aws-lambda-powertools/commons/typeutils';
 
+/**
+ * A data structure that holds a value and its byte size.
+ *
+ * @internal
+ */
 export class SizedItem<V> {
   public value: V;
   public logLevel: number;
@@ -15,16 +20,31 @@ export class SizedItem<V> {
   }
 }
 
+/**
+ * A set that tracks its current byte size.
+ *
+ * @internal
+ */
 export class SizedSet<V> extends Set<SizedItem<V>> {
   public currentBytesSize = 0;
   public hasEvictedLog = false;
 
+  /**
+   * Adds an item to the set and updates the current byte size.
+   *
+   * @param item - The item to add
+   */
   add(item: SizedItem<V>): this {
     this.currentBytesSize += item.byteSize;
     super.add(item);
     return this;
   }
 
+  /**
+   * Deletes an item from the set and updates the current byte size.
+   *
+   * @param item - The item to delete
+   */
   delete(item: SizedItem<V>): boolean {
     const wasDeleted = super.delete(item);
     if (wasDeleted) {
@@ -33,11 +53,17 @@ export class SizedSet<V> extends Set<SizedItem<V>> {
     return wasDeleted;
   }
 
+  /**
+   * Clears all items from the set and resets the current byte size.
+   */
   clear(): void {
     super.clear();
     this.currentBytesSize = 0;
   }
 
+  /**
+   * Removes the first item from the set and returns it.
+   */
   shift(): SizedItem<V> | undefined {
     const firstElement = this.values().next().value;
     if (firstElement) {
@@ -47,6 +73,11 @@ export class SizedSet<V> extends Set<SizedItem<V>> {
   }
 }
 
+/**
+ * A ring buffer that stores logs in a circular manner.
+ *
+ * @internal
+ */
 export class CircularMap<V> extends Map<string, SizedSet<V>> {
   readonly #maxBytesSize: number;
   readonly #onBufferOverflow?: () => void;
@@ -63,6 +94,13 @@ export class CircularMap<V> extends Map<string, SizedSet<V>> {
     this.#onBufferOverflow = onBufferOverflow;
   }
 
+  /**
+   * Adds an item to the buffer, evicting older items if necessary.
+   *
+   * @param key - The key to associate with the item
+   * @param value - The item to add
+   * @param logLevel - The log level of the item
+   */
   setItem(key: string, value: V, logLevel: number): this {
     const item = new SizedItem<V>(value, logLevel);
 
@@ -84,13 +122,19 @@ export class CircularMap<V> extends Map<string, SizedSet<V>> {
     return this;
   }
 
-  readonly #deleteFromBufferUntilSizeIsLessThanMax = (
+  /**
+   * Deletes an item from the buffer.
+   *
+   * @param key - The key associated with the item
+   * @param value - The item to delete
+   */
+  #deleteFromBufferUntilSizeIsLessThanMax(
     buffer: SizedSet<V>,
     item: SizedItem<V>
-  ) => {
+  ) {
     while (buffer.currentBytesSize + item.byteSize >= this.#maxBytesSize) {
       buffer.shift();
       buffer.hasEvictedLog = true;
     }
-  };
+  }
 }