+++ packages/logger/src/types/Logger.ts
@@ -54,6 +54,11 @@ type InjectLambdaContextOptions = {
    * @default `false`
    */
   flushBufferOnUncaughtError?: boolean;
+
+  /**
+   * The path to the correlation ID in the event object, used to extract the correlation ID from the event object and add it to the log attributes.
+   */
+  correlationIdPath?: string;
 };
 
 /**
@@ -201,15 +206,27 @@ type LogBufferOption = {
      */
     flushOnErrorLog?: boolean;
     /**
-     * The threshold to buffer logs. Logs with a level below
-     * this threshold will be buffered
+     * The threshold to buffer logs. Logs with a level more severe than this will be logged immediately.
+     * Only 'DEBUG', 'INFO', 'WARN' or their lowercase variants are allowed.
      * @default `DEBUG`
      */
-    bufferAtVerbosity?: Omit<
-      LogLevel,
-      'ERROR' | 'error' | 'CRITICAL' | 'critical' | 'SILENT' | 'silent'
-    >;
+    bufferAtVerbosity?:
+      | Extract<
+          (typeof LogLevelList)[keyof typeof LogLevelList],
+          'DEBUG' | 'INFO' | 'WARN'
+        >
+      | Extract<
+          Lowercase<(typeof LogLevelList)[keyof typeof LogLevelList]>,
+          'debug' | 'info' | 'warn'
+        >;
   };
+};
+
+type CorrelationIdOption = {
+  /**
+   * The search function for the correlation ID.
+   */
+  correlationIdSearchFn?: (expression: string, data: unknown) => unknown;
 };
 
 /**
@@ -218,7 +235,8 @@ type LogBufferOption = {
 type ConstructorOptions = BaseConstructorOptions &
   (PersistentKeysOption | DeprecatedPersistentKeysOption) &
   (LogFormatterOption | LogRecordOrderOption) &
-  LogBufferOption;
+  LogBufferOption &
+  CorrelationIdOption;
 
 type LogItemMessage = string | LogAttributesWithMessage;
 type LogItemExtraInput = [Error | string] | LogAttributes[];
@@ -247,6 +265,7 @@ type LoggerInterface = {
   removeKeys(keys?: string[]): void;
   removePersistentLogAttributes(keys?: string[]): void;
   resetKeys(): void;
+  setCorrelationId(value: unknown, correlationIdPath?: string): void;
   setLogLevel(logLevel: LogLevel): void;
   setPersistentLogAttributes(attributes?: LogAttributes): void;
   shouldLogEvent(overwriteValue?: boolean): boolean;
@@ -255,14 +274,14 @@ type LoggerInterface = {
 };
 
 export type {
+  ConstructorOptions,
+  CustomJsonReplacerFn,
   Environment,
+  InjectLambdaContextOptions,
   LogAttributes,
-  LogLevel,
   LogFunction,
   LoggerInterface,
-  LogItemMessage,
   LogItemExtraInput,
-  ConstructorOptions,
-  InjectLambdaContextOptions,
-  CustomJsonReplacerFn,
+  LogItemMessage,
+  LogLevel,
 };