+++ packages/logger/src/types/index.ts
@@ -1,12 +1,12 @@
 export type {
-  Environment,
-  LogAttributes,
-  LogLevel,
-  LogItemMessage,
-  LogItemExtraInput,
   ConstructorOptions,
-  InjectLambdaContextOptions,
   CustomJsonReplacerFn,
+  Environment,
+  InjectLambdaContextOptions,
+  LogAttributes,
   LoggerInterface,
+  LogItemExtraInput,
+  LogItemMessage,
+  LogLevel,
 } from './Logger.js';
 export type { UnformattedAttributes } from './logKeys.js';